+++ packages/experimental/Shimmer/src/Shimmer.styling.ts
@@ -5,37 +5,30 @@ import { shimmerName, ShimmerProps, ShimmerSlotProps, ShimmerTokens } from './Sh
 export const stylingSettings: UseStylingOptions<ShimmerProps, ShimmerSlotProps, ShimmerTokens> = {
   tokens: [
     () => ({
-      toValue: 30,
-      duration: 7000,
-      delay: 0,
-      gradientTintColor: 'white', // Baseline component is the light mode color
-      shimmerTintColor: '#E1E1E1', // Baseline component is the light mode color
-      width: 200,
-      height: 100,
       angle: 0,
-      gradientOpacity: 0.7,
+      delay: 0,
+      duration: 7000,
+      shimmerColor: '#E1E1E1', // Baseline component is the light mode color
+      shimmerColorOpacity: 1,
+      shimmerWaveColor: 'white', // Baseline component is the light mode color
+      shimmerWaveColorOpacity: 1,
     }),
     shimmerName,
   ],
   slotProps: {
     root: buildProps(
       (tokens: ShimmerTokens, theme: Theme) => ({
-        toValue: tokens.toValue,
-        duration: tokens.duration,
-        delay: tokens.delay,
-        gradientTintColor: getCurrentAppearance(theme.host.appearance, 'light') === 'light' ? 'white' : 'black',
-        shimmerTintColor: getCurrentAppearance(theme.host.appearance, 'light') === 'light' ? '#E1E1E1' : '#404040',
-        width: tokens.width,
-        height: tokens.height,
-        angle: tokens.angle,
-        gradientOpacity: tokens.gradientOpacity,
         accessibilityRole: 'progressbar',
         accessible: true,
+        angle: tokens.angle,
+        delay: tokens.delay,
+        duration: tokens.duration,
+        shimmerColor: getCurrentAppearance(theme.host.appearance, 'light') === 'light' ? '#E1E1E1' : '#404040',
+        shimmerColorOpacity: tokens.shimmerColorOpacity,
+        shimmerWaveColor: getCurrentAppearance(theme.host.appearance, 'light') === 'light' ? 'white' : 'black',
+        shimmerWaveColorOpacity: tokens.shimmerWaveColorOpacity,
       }),
-      ['gradientTintColor', 'toValue', 'duration', 'delay', 'shimmerTintColor', 'width', 'height', 'angle', 'gradientOpacity'],
+      ['angle', 'delay', 'duration', 'shimmerColor', 'shimmerColorOpacity', 'shimmerWaveColor', 'shimmerWaveColorOpacity'],
     ),
-    image: buildProps((_tokens: ShimmerTokens) => ({
-      href: null,
-    })),
   },
 };