+++ packages/deprecated/foundation-settings/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @uifabricshared/foundation-settings
 
-This log was last generated on Wed, 04 Aug 2021 06:26:25 GMT and should not be manually modified.
+This log was last generated on Sat, 07 Aug 2021 00:40:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.10.1
+
+Sat, 07 Aug 2021 00:40:04 GMT
+
+### Patches
+
+- enable usePressableState with stock Pressable component (jasonmo@microsoft.com)
 
 ## 0.10.0