+++ packages/deprecated/theming-react-native/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @uifabricshared/theming-react-native
 
-This log was last generated on Fri, 06 Aug 2021 20:26:18 GMT and should not be manually modified.
+This log was last generated on Mon, 09 Aug 2021 22:49:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.6
+
+Mon, 09 Aug 2021 22:49:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.7.8 (ruaraki@microsoft.com)
+
+## 0.12.5
+
+Mon, 09 Aug 2021 19:16:08 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.7.7 (ruaraki@microsoft.com)
+
+## 0.12.4
+
+Sat, 07 Aug 2021 00:40:04 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.7.6 (jasonmo@microsoft.com)
+
+## 0.12.3
+
+Fri, 06 Aug 2021 23:18:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.7.5 (ruaraki@microsoft.com)
 
 ## 0.12.2